+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.2] - 2026-06-25
+
+### Removed
+
+- Legacy code relating to old V1 endpoints
+
 ## [2.9.1] - 2026-06-25
 
 ### Changed
@@ -63,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   await passlock.authorizePasskeyRegistration({
     rpId: 'example.com',
+    rpName: 'Example App',
     ...
   });